more portable types

+++ hash.c
@@ -54,7 +54,7 @@ EXPORT unsigned char *make_hash(metafile_t *m)
 	                                   the read buffer */
 	SHA_CTX c;                      /* SHA1 hashing context */
 #ifndef NO_HASH_CHECK
-	unsigned long long counter = 0; /* number of bytes hashed
+	unsigned long counter = 0;      /* number of bytes hashed
 	                                   should match size when done */
 #endif
 
@@ -130,7 +130,7 @@ EXPORT unsigned char *make_hash(metafile_t *m)
 #ifndef NO_HASH_CHECK
 	counter += r;
 	if (counter != m->size) {
-		fprintf(stderr, "Counted %llu bytes, but hashed %llu bytes. "
+		fprintf(stderr, "Counted %lu bytes, but hashed %lu bytes. "
 				"Something is wrong...\n", m->size, counter);
 		exit(EXIT_FAILURE);
 	}
